small refatcor
