+++ src/components/modules/questionSection/answerArea/multiChoice/styles.js
@@ -1,0 +1,30 @@
+export default {
+  myStyle: {
+    borderRadius: 5,
+    display: 'flex',
+    flexDirection: 'flex-wrap',
+    justifyContent: 'center',
+    color: 'yellow',
+    backgroundColor: 'blue',
+    alignItems: 'center',
+    width: '100%',
+    margin: 10,
+    flexGrow: 1,
+  },
+  selectedStyle: {
+    backgroundColor: 'lightGreen',
+    color: 'black',
+  },
+  buttonRow: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    flexGrow: 1,
+    alignItems: 'stretch',
+  },
+  multiContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    flexGrow: 1,
+  },
+}
